requirements.txt
falcon
waitress
passlib
bcrypt
requests
pymongo
marshmallow-dataclass
dataclasses-json
pyjwt